refactor: split cli/run_cmd.py into run_cmd/ package
