+++ initrunner/cli/run_cmd/__init__.py
@@ -1,0 +1,5 @@
+"""Run command: unified dispatcher for agent, team, flow, and ephemeral modes."""
+
+from initrunner.cli.run_cmd._command import run
+
+__all__ = ["run"]
