+++ tests/test_run_report.py
@@ -74,8 +74,8 @@ class TestExportReportCLI:
         result_obj = _successful_run_result()
 
         with (
-            patch("initrunner.cli.run_cmd.resolve_run_target", _passthrough_run_target),
-            patch("initrunner.cli.run_cmd.preflight_validate_or_exit", lambda _p: None),
+            patch("initrunner.cli.run_cmd._command.resolve_run_target", _passthrough_run_target),
+            patch("initrunner.cli.run_cmd._command.preflight_validate_or_exit", lambda _p: None),
             patch("initrunner.cli._run_agent.command_context", _mock_command_context()),
         ):
             with patch("initrunner.runner.run_single") as mock_run:
@@ -106,8 +106,8 @@ class TestExportReportCLI:
         result_obj = _successful_run_result()
 
         with (
-            patch("initrunner.cli.run_cmd.resolve_run_target", _passthrough_run_target),
-            patch("initrunner.cli.run_cmd.preflight_validate_or_exit", lambda _p: None),
+            patch("initrunner.cli.run_cmd._command.resolve_run_target", _passthrough_run_target),
+            patch("initrunner.cli.run_cmd._command.preflight_validate_or_exit", lambda _p: None),
             patch("initrunner.cli._run_agent.command_context", _mock_command_context()),
         ):
             with patch("initrunner.runner.run_single") as mock_run:
@@ -138,8 +138,8 @@ class TestExportReportCLI:
         """Unknown --report-template errors before execution."""
         report_file = tmp_path / "report.md"
         with (
-            patch("initrunner.cli.run_cmd.resolve_run_target", _passthrough_run_target),
-            patch("initrunner.cli.run_cmd.preflight_validate_or_exit", lambda _p: None),
+            patch("initrunner.cli.run_cmd._command.resolve_run_target", _passthrough_run_target),
+            patch("initrunner.cli.run_cmd._command.preflight_validate_or_exit", lambda _p: None),
             patch("initrunner.cli._run_agent.command_context", _mock_command_context()),
         ):
             result = runner.invoke(
@@ -162,8 +162,8 @@ class TestExportReportCLI:
     def test_report_template_without_report_errors(self):
         """--report-template without --report must error."""
         with (
-            patch("initrunner.cli.run_cmd.resolve_run_target", _passthrough_run_target),
-            patch("initrunner.cli.run_cmd.preflight_validate_or_exit", lambda _p: None),
+            patch("initrunner.cli.run_cmd._command.resolve_run_target", _passthrough_run_target),
+            patch("initrunner.cli.run_cmd._command.preflight_validate_or_exit", lambda _p: None),
             patch("initrunner.cli._run_agent.command_context", _mock_command_context()),
         ):
             result = runner.invoke(
@@ -187,8 +187,8 @@ class TestExportReportCLI:
         result_obj = _failed_run_result()
 
         with (
-            patch("initrunner.cli.run_cmd.resolve_run_target", _passthrough_run_target),
-            patch("initrunner.cli.run_cmd.preflight_validate_or_exit", lambda _p: None),
+            patch("initrunner.cli.run_cmd._command.resolve_run_target", _passthrough_run_target),
+            patch("initrunner.cli.run_cmd._command.preflight_validate_or_exit", lambda _p: None),
             patch("initrunner.cli._run_agent.command_context", _mock_command_context()),
         ):
             with patch("initrunner.runner.run_single") as mock_run:
